+++ revalidation/style.css
@@ -1,0 +1,22 @@
+dialog.wporg-2fa-revalidate-modal {
+	border-radius: 8px;
+}
+dialog.wporg-2fa-revalidate-modal > h1 {
+	margin: unset;
+	margin-bottom: 1em;
+	text-align: center;
+}
+dialog.wporg-2fa-revalidate-modal > iframe {
+	width: 100%;
+	height: 330px; /* Room for errors. */
+	border: none;
+}
+/* Close button. */
+dialog.wporg-2fa-revalidate-modal > button {
+	position: absolute;
+	top: 0;
+	right: 0;
+	background: none;
+	border: none;
+	padding: 15px;
+}
